backend/requirements.txt
tensorflow==2.19.0
keras==3.10.0
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
pillow==10.4.0
opencv-python==4.10.0.84
flask==3.0.3
fastapi==0.115.0
uvicorn==0.30.3
boto3==1.35.10
awscli==1.34.10
h5py==3.11.0
protobuf==4.25.3